Signal semantics redesign + Items 4, 6: periodic cost decouple, get_weights
Signal redesign (BREAKING):
- signal=0 now means "go flat" (was "hold")
- NaN means "hold" (was silently converted to 0)
- Continuous signals: value = exposure fraction of max target
- allow_short=False + negative signal: warn + skip (was: reinterpret)
- signal_transform config field for z-score mapping
- Pipeline preserves NaN (stops fillna(0))

Other changes:
- PeriodicCostModel decoupled from MarginConfig (FundingRate works alone)
- Portfolio.get_weights() convenience method
- Allocator signal type int → float
- All tests migrated from signal=0 to signal=NaN for hold semantics

102 tests pass.

+++ src/aiphaforge/capital_allocator.py
@@ -27,7 +27,7 @@ class BaseCapitalAllocator:
 
     def allocate(
         self,
-        signals: Dict[str, int],
+        signals: Dict[str, float],
         prices: Dict[str, float],
         portfolio: Any,
         config: Any,
@@ -56,7 +56,7 @@ class EqualWeightAllocator(BaseCapitalAllocator):
 
     def allocate(
         self,
-        signals: Dict[str, int],
+        signals: Dict[str, float],
         prices: Dict[str, float],
         portfolio: Any,
         config: Any,
@@ -97,7 +97,7 @@ class FixedWeightAllocator(BaseCapitalAllocator):
 
     def allocate(
         self,
-        signals: Dict[str, int],
+        signals: Dict[str, float],
         prices: Dict[str, float],
         portfolio: Any,
         config: Any,
@@ -136,7 +136,7 @@ class ProRataAllocator(BaseCapitalAllocator):
 
     def allocate(
         self,
-        signals: Dict[str, int],
+        signals: Dict[str, float],
         prices: Dict[str, float],
         portfolio: Any,
         config: Any,
@@ -168,7 +168,7 @@ class MarginAllocator(BaseCapitalAllocator):
 
     def allocate(
         self,
-        signals: Dict[str, int],
+        signals: Dict[str, float],
         prices: Dict[str, float],
         portfolio: Any,
         config: Any,
